requirements.txt
# --- FastAPI Core ---
fastapi==0.111.0
uvicorn[standard]==0.29.0

# --- OpenAI & Embeddings ---
openai==1.30.1
tiktoken==0.7.0

# --- Vector Store (RAG) ---
chromadb==0.4.24

# --- JSON/YAML support, if needed ---
python-dotenv==1.0.1

# --- Optional (for logging/debugging) ---
ipython==8.24.0
